exec: improve aggregator tests and benchmarks and fix bug
All aggregate tests test both hash and ordered aggregator now.

The benchmarks weren't working right - they vastly overestimated the
speed of the operator, unfortunately, because their reset methods didn't
work.

This commit splits the benchmark into one for ints and one for decimals,
to showcase how slow decimals are in comparison. It also fixes the reset
methods, and coalesces the hash and ordered aggregator benchmarks into
one.

```
BenchmarkAggregator/ANY_NOT_NULL/hash/Int64/groupSize=1/numInputBatches=64-8                 500           2819938 ns/op         185.92 MB/s
BenchmarkAggregator/ANY_NOT_NULL/hash/Int64/groupSize=2/numInputBatches=64-8                 500           2809348 ns/op         186.62 MB/s
BenchmarkAggregator/ANY_NOT_NULL/hash/Int64/groupSize=512/numInputBatches=64-8               300           4777525 ns/op         109.74 MB/s
BenchmarkAggregator/ANY_NOT_NULL/hash/Int64/groupSize=1024/numInputBatches=64-8              200           6364571 ns/op          82.38 MB/s
BenchmarkAggregator/ANY_NOT_NULL/hash/Decimal/groupSize=1/numInputBatches=64-8               300           3984492 ns/op         131.58 MB/s
BenchmarkAggregator/ANY_NOT_NULL/hash/Decimal/groupSize=2/numInputBatches=64-8               300           3583325 ns/op         146.31 MB/s
BenchmarkAggregator/ANY_NOT_NULL/hash/Decimal/groupSize=512/numInputBatches=64-8             200           5713770 ns/op          91.76 MB/s
BenchmarkAggregator/ANY_NOT_NULL/hash/Decimal/groupSize=1024/numInputBatches=64-8            200           6946676 ns/op          75.47 MB/s
BenchmarkAggregator/ANY_NOT_NULL/ordered/Int64/groupSize=1/numInputBatches=64-8             5000            247468 ns/op        2118.61 MB/s
BenchmarkAggregator/ANY_NOT_NULL/ordered/Int64/groupSize=2/numInputBatches=64-8             5000            200777 ns/op        2611.28 MB/s
BenchmarkAggregator/ANY_NOT_NULL/ordered/Int64/groupSize=512/numInputBatches=64-8          10000            159597 ns/op        3285.07 MB/s
BenchmarkAggregator/ANY_NOT_NULL/ordered/Int64/groupSize=1024/numInputBatches=64-8         10000            157453 ns/op        3329.79 MB/s
BenchmarkAggregator/ANY_NOT_NULL/ordered/Decimal/groupSize=1/numInputBatches=64-8           2000            594968 ns/op         881.20 MB/s
BenchmarkAggregator/ANY_NOT_NULL/ordered/Decimal/groupSize=2/numInputBatches=64-8           3000            335296 ns/op        1563.66 MB/s
BenchmarkAggregator/ANY_NOT_NULL/ordered/Decimal/groupSize=512/numInputBatches=64-8        10000            158746 ns/op        3302.67 MB/s
BenchmarkAggregator/ANY_NOT_NULL/ordered/Decimal/groupSize=1024/numInputBatches=64-8       10000            157495 ns/op        3328.92 MB/s
BenchmarkAggregator/COUNT_ROWS/hash/Int64/groupSize=1/numInputBatches=64-8                   500           2688660 ns/op         195.00 MB/s
BenchmarkAggregator/COUNT_ROWS/hash/Int64/groupSize=2/numInputBatches=64-8                   500           2780675 ns/op         188.55 MB/s
BenchmarkAggregator/COUNT_ROWS/hash/Int64/groupSize=512/numInputBatches=64-8                 300           4715001 ns/op         111.20 MB/s
BenchmarkAggregator/COUNT_ROWS/hash/Int64/groupSize=1024/numInputBatches=64-8                200           6511204 ns/op          80.52 MB/s
BenchmarkAggregator/COUNT_ROWS/hash/Decimal/groupSize=1/numInputBatches=64-8                 500           2658724 ns/op         197.20 MB/s
BenchmarkAggregator/COUNT_ROWS/hash/Decimal/groupSize=2/numInputBatches=64-8                 500           2757335 ns/op         190.14 MB/s
BenchmarkAggregator/COUNT_ROWS/hash/Decimal/groupSize=512/numInputBatches=64-8               300           4861616 ns/op         107.84 MB/s
BenchmarkAggregator/COUNT_ROWS/hash/Decimal/groupSize=1024/numInputBatches=64-8              200           6405883 ns/op          81.84 MB/s
BenchmarkAggregator/COUNT_ROWS/ordered/Int64/groupSize=1/numInputBatches=64-8               5000            300708 ns/op        1743.51 MB/s
BenchmarkAggregator/COUNT_ROWS/ordered/Int64/groupSize=2/numInputBatches=64-8               5000            275161 ns/op        1905.38 MB/s
BenchmarkAggregator/COUNT_ROWS/ordered/Int64/groupSize=512/numInputBatches=64-8             5000            260265 ns/op        2014.43 MB/s
BenchmarkAggregator/COUNT_ROWS/ordered/Int64/groupSize=1024/numInputBatches=64-8            5000            264332 ns/op        1983.44 MB/s
BenchmarkAggregator/COUNT_ROWS/ordered/Decimal/groupSize=1/numInputBatches=64-8             5000            298114 ns/op        1758.68 MB/s
BenchmarkAggregator/COUNT_ROWS/ordered/Decimal/groupSize=2/numInputBatches=64-8             5000            270498 ns/op        1938.23 MB/s
BenchmarkAggregator/COUNT_ROWS/ordered/Decimal/groupSize=512/numInputBatches=64-8           5000            268750 ns/op        1950.84 MB/s
BenchmarkAggregator/COUNT_ROWS/ordered/Decimal/groupSize=1024/numInputBatches=64-8          5000            257533 ns/op        2035.80 MB/s
BenchmarkAggregator/SUM/hash/Int64/groupSize=1/numInputBatches=64-8                          500           2836210 ns/op         184.86 MB/s
BenchmarkAggregator/SUM/hash/Int64/groupSize=2/numInputBatches=64-8                          500           2932260 ns/op         178.80 MB/s
BenchmarkAggregator/SUM/hash/Int64/groupSize=512/numInputBatches=64-8                        300           5039520 ns/op         104.04 MB/s
BenchmarkAggregator/SUM/hash/Int64/groupSize=1024/numInputBatches=64-8                       200           6537244 ns/op          80.20 MB/s
BenchmarkAggregator/SUM/hash/Decimal/groupSize=1/numInputBatches=64-8                        100          14244380 ns/op          36.81 MB/s
BenchmarkAggregator/SUM/hash/Decimal/groupSize=2/numInputBatches=64-8                        100          13545369 ns/op          38.71 MB/s
BenchmarkAggregator/SUM/hash/Decimal/groupSize=512/numInputBatches=64-8                       20          52605020 ns/op           9.97 MB/s
BenchmarkAggregator/SUM/hash/Decimal/groupSize=1024/numInputBatches=64-8                      20          56025125 ns/op           9.36 MB/s
BenchmarkAggregator/SUM/ordered/Int64/groupSize=1/numInputBatches=64-8                      5000            297225 ns/op        1763.94 MB/s
BenchmarkAggregator/SUM/ordered/Int64/groupSize=2/numInputBatches=64-8                      5000            279924 ns/op        1872.96 MB/s
BenchmarkAggregator/SUM/ordered/Int64/groupSize=512/numInputBatches=64-8                    5000            261397 ns/op        2005.72 MB/s
BenchmarkAggregator/SUM/ordered/Int64/groupSize=1024/numInputBatches=64-8                   5000            268299 ns/op        1954.11 MB/s
BenchmarkAggregator/SUM/ordered/Decimal/groupSize=1/numInputBatches=64-8                     100          10972373 ns/op          47.78 MB/s
BenchmarkAggregator/SUM/ordered/Decimal/groupSize=2/numInputBatches=64-8                     100          10065695 ns/op          52.09 MB/s
BenchmarkAggregator/SUM/ordered/Decimal/groupSize=512/numInputBatches=64-8                    30          46431390 ns/op          11.29 MB/s
BenchmarkAggregator/SUM/ordered/Decimal/groupSize=1024/numInputBatches=64-8                   30          48812805 ns/op          10.74 MB/s
```

Release note: None

+++ pkg/sql/exec/coalescer_test.go
@@ -85,8 +85,8 @@ func BenchmarkCoalescer(b *testing.B) {
 
 	for colIdx := 0; colIdx < nCols; colIdx++ {
 		col := batch.ColVec(colIdx).Int64()
-		for i := uint16(0); i < coldata.BatchSize; i++ {
-			col[i] = int64(i)
+		for i := int64(0); i < coldata.BatchSize; i++ {
+			col[i] = i
 		}
 	}
